+++ CHANGELOG.md
@@ -37,6 +37,8 @@ however, insignificant breaking changes does not guarantee a major version bump,
 - A reworked interface for `?snippet` and `?alias`.
   - Add an `?snippet raw <name>` command for viewing the raw content of a snippet (escaped markdown).
 - The placeholder channel for the streaming status changed to https://www.twitch.tv/discordmodmail/.
+- Removed unclear `rm` alias for some `remove` commands.
+- Paginate `?config options`.
 
 ### Fixes